Refactoring Restaurant Owners' Methods
